Compile farm fest 	Debugged and verified compilation on all platforms. 	Numerous changes to bsconf for BSDs and Macs. 	Link bvts with libsupc++ if no shared lib is built. 	Update bvt05 output to work on most platforms 		Some platforms still offer no support for %zd or %td; I'll just 		ignore them, since there is no easy workaround and it's a test. 	Changed sistream str_to_num to spec by parameter, not rv 		This is needed to keep supporting 2.95 (on Alpha) 	Redefine UINTPTR_MAX on Solaris which defines it empty (?!?) 	Remove -Werror from flags. 	Renamed STANDALONE to NOLIBSTDCPP 	Always build throwing_malloc and free_nullok (used by CBacktrace) 	Reworked vformat again, to support ancient vsnprintfs that return 		-1 instead of needed space (MacOS) 	Fixed MMX passign for 4,float; wrong arguments copied. 	__builtin_popcount is in gcc 4+ only. 	Don't include <locale> from uspecial; always use ctype.h. 	Added ostream::SetPos to simplify cases where no overflow can occur. 	Restrict SSE load_identity to linux, since BSD and Mac gccs 		don't know how to load 1.0f into an xmm register (?)

+++ sostream.h
@@ -42,7 +42,7 @@ public:
     inline void			iwrite (unsigned long long v)	{ iformat (v); }
 #endif
     inline size_type		max_size (void) const		{ return (m_Buffer.max_size()); }
-    inline void			put (char c)			{ iwrite (c); }
+    inline void			put (char c)			{ iwrite (uint8_t(c)); }
     int				vformat (const char* fmt, va_list args);
     int				format (const char* fmt, ...) __attribute__((__format__(__printf__, 2, 3)));
     inline void			set_base (uint16_t b)		{ m_Base = b; }
